% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{AP_sim2comp}
\alias{AP_sim2comp}
\title{Structure Similarity between two new compound}
\usage{
AP_sim2comp(apset_compA, apset_compB)
}
\arguments{
\item{apset_compA}{APset object containing the Atom Pair descriptors of new compound}

\item{apset_compB}{APset object containing the Atom Pair descriptors of new compound}
}
\value{
data.frame containing the similarity score between two selected compounds
}
\description{
Compute the structure similarity score between the selected compound and the compounds of CMAP
}
