% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{AP_simCMAP}
\alias{AP_simCMAP}
\title{Structure Similarity compound-CMAPset}
\usage{
AP_simCMAP(apset_comp, apsetCMAP)
}
\arguments{
\item{apset_comp}{APset object containing the Atom Pair descriptors of new compound(s)}

\item{apsetCMAP}{APset object containing all Atom Pair descriptors of CMAP compounds}
}
\value{
data.frame (Nx1) with similarity score between new_compound and apsetCMAP
}
\description{
Compute the structure similarity score between the selected compound and the compounds of CMAP
}
