% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugInfo_class.R
\docType{class}
\name{DrugInfoCMAP-class}
\alias{DrugInfoCMAP-class}
\alias{DrugInfoCMAP}
\title{An S4 class to represent a Drug information set}
\description{
This class represents Drug information set in a S4 DrugInfoCMAP object.
}
\section{Fields}{

\describe{
\item{\code{APsdf}}{APset. It should contains the atom pairs for each CMAP drug}

\item{\code{GeneExpr}}{data.frame. It should contain the gene expression profile for each CMAP drug}

\item{\code{pmid}}{list. It should contain the set of PMID in which the compound name occured.}

\item{\code{target}}{character. It is an array containing the targets of each CMAP drug}

\item{\code{path}}{character. It is an array containing the path of each CMAP drug}

\item{\code{maps}}{character. It is an array containing the maps of each CMAP drug}

\item{\code{ATC}}{data frame. It is a data.frame containing the ATC code divided by each drug classification level.}
}}

