% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{geneExpr_upload}
\alias{geneExpr_upload}
\title{Gene Expression data upload when it is available}
\usage{
geneExpr_upload(DrugInfo)
}
\arguments{
\item{DrugInfo}{list containing the info of new selected compounds}
}
\value{
DrugInfo_list the input list will be updated adding the Gene expression profiles if it is available.
}
\description{
Gene expression file upload in data.frame type object
}
