% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_pmid}
\alias{get_pmid}
\title{get PMID file from PubMed}
\usage{
get_pmid(DrugInfo_list)
}
\arguments{
\item{DrugInfo_list}{list containing the drug information contained in configuration file}
}
\value{
DrugInfo_list the input list will be updated adding the PMIDs list of new compounds
}
\description{
get PMID file from PubMed
}
