% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_sdf}
\alias{get_sdf}
\title{Get sdf file from PubChem and conversion of the sdf in Atom Pair descriptors}
\usage{
get_sdf(DrugInfo_list, result_folder)
}
\arguments{
\item{DrugInfo_list}{list containing the drug information contained in configuration file}

\item{result_folder}{character folder name that will be contained all output files (see makeResultsDirectory function)}
}
\value{
DrugInfo_list the input list will be updated adding the Atom pair descriptor in APsdf key
}
\description{
Connection tu PubChem to download the sdf file
}
