% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{GetAllSimilarityScores}
\alias{GetAllSimilarityScores}
\title{All similarity score computation among new compounds}
\usage{
GetAllSimilarityScores(newDrugInfo, new_comp_A, new_comp_B)
}
\arguments{
\item{newDrugInfo}{list containing the drug information of the new selected compounds}

\item{new_comp_A}{the name of one drug}

\item{new_comp_B}{the name of another drug}
}
\value{
list object containing the similairty scores based on all chemical features among two new compounds
}
\description{
Compute all type similarity scores among new compounds
}
