% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{GetAllSimilarityScoresCMAP}
\alias{GetAllSimilarityScoresCMAP}
\title{All similarity score computation between two new compound and connecitivity map compounds}
\usage{
GetAllSimilarityScoresCMAP(CMAPData, newDrugInfo)
}
\arguments{
\item{CMAPData}{Drug Info CMAP Class object containing the CMAP compounds information}

\item{newDrugInfo}{list containing the drug information of the new selected compounds}
}
\value{
list object containing the similairty scores based on all chemical features between new compounds and CMAP compounds
}
\description{
Compute the KEGG data score between the selected compound and the compounds of CMAP
}
