% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{GetATCcodeScore}
\alias{GetATCcodeScore}
\title{ATC code score similarity between a new compound and a set of ATC codes}
\usage{
GetATCcodeScore(newDrugInfo_comp, ATCcode_df)
}
\arguments{
\item{newDrugInfo_comp}{list object containing the information of a new compound}

\item{ATCcode_df}{data.frame containing the ATC codes for one or more compounds}
}
\value{
data.frame containing the compound pair similarity scores between the ATC code of the new compound and the ATC codes provided in ATCcode_df
}
\description{
Compute the ATC score between a selected compound and one or more compounds
}
