% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{GetGeneExprScore}
\alias{GetGeneExprScore}
\title{Gene Expression similarity between two new compound and connecitivity map compounds}
\usage{
GetGeneExprScore(GeneExprNewDrug, GeneExprCMAP)
}
\arguments{
\item{GeneExprNewDrug}{data.frame containing the gene expression of a new compound}

\item{GeneExprCMAP}{data.frame containing the gene expression profiles of CMAP}
}
\value{
data.frame object containing the gene expression correlation between the new compound and all CMAP compounds
}
\description{
Compute the Correlation score between the selected compound and the compounds of CMAP
}
