% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityScoreFunctions.R
\name{GetLitteratureCo_occ}
\alias{GetLitteratureCo_occ}
\title{co-occurrence similarity between two new compound and connectivity map compounds}
\usage{
GetLitteratureCo_occ(new_comp_name, PMID_newComps, PMID_cmap)
}
\arguments{
\item{new_comp_name}{a character object containing the name associated to the new compound}

\item{PMID_newComps}{character array containing the PMIDs associated to the new compound}

\item{PMID_cmap}{list object containing the PMID associated to each compound of CMAP}
}
\value{
data frame containing the compound pair similarity scores between the new compound(s) and all CMAP compounds
}
\description{
Compute the co-occurrence score between the selected compound and the compounds of CMAP
}
