% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugSimilarityPredictiction.R
\name{LoadModelData}
\alias{LoadModelData}
\title{Predictive Model data and CMAP similarity data}
\usage{
LoadModelData()
}
\arguments{
\item{SimilairtyScoreList}{object obtained by GetAllSimilarityScoresCMAP function}
}
\value{
list object containing the Predicitve Model and a data.frame containing all CMAP chemical similarity scores and the phenotypic predicted score
}
\description{
Load Predictive Model data and CMAP compound pair similarity
}
