% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{LoadPackage}
\alias{LoadPackage}
\title{Import packages}
\usage{
LoadPackage(package = FALSE)
}
\arguments{
\item{package}{specific package to be loaded}
}
\description{
Import a new package and default packages
}
