% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{makeResultsDirectory}
\alias{makeResultsDirectory}
\title{Make result folder}
\usage{
makeResultsDirectory(DesiredFolderName)
}
\arguments{
\item{DesiredFolderName}{character. Name of folder will contain the results}
}
\value{
the output_folder path that will contain the date as suffix.
}
\description{
Creation of new folder
}
