% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{OrderCompound}
\alias{OrderCompound}
\title{Order}
\usage{
OrderCompound(CompA, CompB)
}
\arguments{
\item{CompA}{character. One or a list of compounds.}

\item{CompB}{character. List of compounds}
}
\value{
comp_pairs
}
\description{
Gene expression file upload in data.frame type object
}
