% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugSimilarityPredictiction.R
\name{PredictNewDrugSimilarity}
\alias{PredictNewDrugSimilarity}
\title{Predicition of new drug similarity scores}
\usage{
PredictNewDrugSimilarity(SimilairtyScoreList, ModelData)
}
\arguments{
\item{SimilairtyScoreList}{object obtained by GetAllSimilarityScoresCMAP function}
}
\value{
InputDataAndPredicted data.frame object that contain the predicted score and the input data passed to the Predictive Model.
}
\description{
Predicition of new drug similarity scores
}
