% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{readDrugInfo}
\alias{readDrugInfo}
\title{Drug Information file reading}
\usage{
readDrugInfo(filename_info)
}
\arguments{
\item{filename_info}{name of the file containing drug information provided as input by the user (e.g. data_example.conf)}
}
\value{
DrugInfo_list object containing Trade/Common name, CID,	Target,	Map,	Path,	ATC (when there are multiple items you have to separate by a double pipe '||')
}
\description{
Read the file containing drug information provided as input by the user
}
