% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{UploadAllData}
\alias{UploadAllData}
\title{Conncectivity map (CMAP) data and new chemical information}
\usage{
UploadAllData(filename_info, OutputFolderName)
}
\arguments{
\item{filename_info}{name of the file containing drug information provided as input by user}

\item{OutputFolderName}{character desired name of the output folder that will contain the results.}
}
\value{
list object containing two other listes:

CMAPData DrugInfoCMAP class object contaning the CMAP data information

Drug_info_list list object containing the information of the new compounds provided in the configuration file (see readDrugInfo function)
}
\description{
Upload of Conncectivity map data in a DrugInfoCMAP class object and upload of the information of desired compounds
}
